protc language tooling (aka protcheck)
First pass a protc language tooling via get-info and #:info integration.

For now we will try to use the drracket:toolbar-buttons key to implement
protocol checking, but we may ultimately need to implement a full tool@
for drracket to make the checking visible. The skeleton code for that is
in the next commit.