curation syntax added protc:vary for annotating multiple conditions
One common pattern that is used in many methods sections is to report
a series of conditions that are varied for a single subject, often in
order. At the moment I am implementing the bare bones needed to allow
protc:implied-vary to be used as a reply to an existing aspect
annotation that has multiple parameters as children.

It is not entirely clear that this is needed since it is probably
better to represent complex parameter values (e.g. a complex number or
a spatial extent defined by a point and a radius) entirely inside the
parameter* annotation. However, there may also be a need to make it
possible to define the syntax or type of an aspect such that it can
validate that the parameter is has been given matches the aspects
type. This a feature for way down the road.