first part of curation testing cleanup
Added notes to code that is about to be removed, specifically the
keyword related tests which very much do not work due to a prior
confusion about how ~? works.

Also added example of the correct way to use check-exn to test for
expected failures when parsing syntax by using convert-syntax-error
from syntax/macro-testing which is much easier than whatever it was
that I was doing in rrid-metadata.