symbols provided via protc-out are provided independently as a list
Needed to use a combination of datum->syntax where the calling
location of the outer macro is capture and passed so that the scope is
set correctly in combination with a macro to wrap a call to
syntax-local-lift-module-end-declaration to put the call to the
previous macro at the end of the module (aka after all calls to
(provide (protc-out name))).

This isn't quite done but should make it much easier to communicate
with the export machinery now.