protio added dual and chain operators
not entirely sure we need chain at all

dual makes it possible to describe the generic type a single process
that is composed of processes that are duals of eachother i.e. that
they are the same overall process, but they have different names
depending on which participant is primary for the purposes of a
particular binding of aspects, e.g. the allocation of stock-water
will change, and the volume of water-in-the-beaker increases
one would be a water pouring technique and the other would be a
beaker filling technique, they do not have to be duals of eachother
but when they are we need a clear way to indicate that part of the "hole"
indicated by the modification of the stock-water in
(>v>' stock-water allocation) and be filled in if we know the exact
amount of (v> water-in-the-beaker volume)