setup.py detect if a wheel install is happening from git
Needed in order to trigger the correct packaging of data_files in the
event that someone or something (usually pip with a git url being
using in a Pipfile) tries to build a wheel directly from git. I try to
avoid this build workflow because it is difficult to control what
files end up being packaged, but it can happen, so we support it.

version bump for release